requirements.txt
torchinfo
ipywidgets
matplotlib
jupyternotify
tqdm